moved ffi components into separate file
