+++ packages/vector_graphics_compiler/lib/src/svg/_path_ops_unsupported.dart
@@ -1,0 +1,9 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// Whether or not tesselation should be used.
+bool get isTesselatorInitialized => false;
+
+/// Initialize the libpathops dynamic library.
+void initializeLibPathOps(String path) {}
